Change license of endianness helpers
Place the new and future versions of the endianness helper files under
the MIT license rather than the GPL. (The existing previous versions
remain unaffected by the license change.)

The reason is the viral character of the GPL, which would disallow the
helpers to be included into non-GPLer projects.

The change allows the helpers to be included within any project, which
is OK because they are simple enough and not a large deal of work. It
won't kill me if they are used in commercial closed-source projects,
too.

Note, however, that not all source files of this project have been put
under the MIT license - only those files which represent the endianness
helper framework.